Add an option to apply the update to target modules that are loaded.
The default is still to abort if some of the target modules are not
loaded, though the abort message now advertises the --partial option.

A consequence of this change is that one now needs to write "cleanup"
into the stage file after reversing an update to clean up the state
related to the update.

Signed-off-by: Tim Abbott <tabbott@mit.edu>