Add .ksplice_call* support to objmanip and inspect.
One should only call Ksplice calls in the case that they are newly
added by the patch.

Also, some of the apply/reverse hooks cannot be safely used if they
can call a patched function (because they are called when the patches
are not applied), and so we check whether this is possible and abort
if so.

Signed-off-by: Tim Abbott <tabbott@mit.edu>