Make sure that one of the other_sect offsets is 0.
This is a hack to work around the fact that our .fixup splitting
algorithm is not correct when there are direct relocations (i.e. not
from __ex_table) into .fixup.  In recent amd64 kernels, there are
direct relocations to the bad_from_user symbol at the start of the
.fixup section in arch/x86/lib/copy_user_64.S, and no indirect
relocations to that same place.

Signed-off-by: Tim Abbott <tabbott@mit.edu>