Add a check for loaded cold-patched modules when reversing an update.
When an update is loaded with the --partial option, one can later load
the modules changed by the update that were not loaded when the update
was applied, so long as one loads cold-patched versions of those
modules.  It is unsafe to reverse such an update while the
cold-patched modules are loaded, so we now prevent doing so.

Signed-off-by: Tim Abbott <tabbott@mit.edu>