Implement a ksplice_match_data_early option.
If a patch's new code references data (for instance a string) that was
present in the old code, but only pointed to from a '.data' section,
Ksplice will fail to run-pre match the update, since we do not use
relocations derived from .data sections when relocating the new code.

To work around this problem, add support for an option that indicates
that the kernel module should run-pre match a given .data object
before trying to relocate the new code, instead of only doing so
afterwards, as an additional sanity check (which is the default
behavior).

Signed-off-by: Nelson Elhage <nelhage@ksplice.com>